requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
pytz==2022.7.1
removebg==0.4
sqlparse==0.4.3
tzdata==2022.7
